requirements.txt
python-telegram-bot==13.15
yt-dlp==2023.7.6
requests==2.31.0
python-dotenv==1.0.0 